social_distance: std::vector<std::atomic<int64_t>>->std::vector<std::atomic<int32_t>>

+++ demos/social_distance_demo/cpp/main.cpp
@@ -166,9 +166,9 @@ struct Context {  // stores all global data for tasks
         std::vector<PersonTrackers> personTracker;
         std::vector<int> minW;
         std::vector<int> maxW;
-        std::vector<std::atomic<int64_t>> lastProcessedIds;
+        std::vector<std::atomic<int32_t>> lastProcessedIds;
 
-        //manual definition of constructor is needed only for creating vector<std::atomic<int64_t>>
+        //manual definition of constructor is needed only for creating vector<std::atomic<int32_t>>
         TrackersContext(std::vector<PersonTrackers>&& personTracker, std::vector<int>&& minW, std::vector<int>&& maxW)
             : personTracker(personTracker), minW(minW), maxW(maxW), lastProcessedIds(personTracker.size()) {
             for (size_t i = 0; i < lastProcessedIds.size(); ++i) {
